Moves `TestSim` to tests
